@@ -6,10 +6,10 @@ import PauseIcon from '@mui/icons-material/Pause';
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
 import ViewInArIcon from '@mui/icons-material/ViewInAr';
 import ImageOutlinedIcon from '@mui/icons-material/ImageOutlined';
-
-import { WebSocketContext } from '../../WebSocket/WebSocket';
-
-const msgpack = require('msgpack-lite');
+import {
+  ViserWebSocketContext,
+  sendWebsocketMessage,
+} from '../../WebSocket/ViserWebSocket';
 
 interface StatusPanelProps {
   sceneTree: object;
@@ -17,17 +17,15 @@ interface StatusPanelProps {
 
 export default function StatusPanel(props: StatusPanelProps) {
   const dispatch = useDispatch();
-  const websocket = React.useContext(WebSocketContext).socket;
+  const viser_websocket = React.useContext(ViserWebSocketContext);
   const isTraining = useSelector((state) => state.renderingState.isTraining);
   const sceneTree = props.sceneTree;
 
   const isWebsocketConnected = useSelector(
     (state) => state.websocketState.isConnected,
   );
+  const step = useSelector((state) => state.renderingState.step);
   const eval_res = useSelector((state) => state.renderingState.eval_res);
-  const vis_train_ratio = useSelector(
-    (state) => state.renderingState.vis_train_ratio,
-  );
   const camera_choice = useSelector(
     (state) => state.renderingState.camera_choice,
   );
@@ -51,6 +49,7 @@ export default function StatusPanel(props: StatusPanelProps) {
         obj.visible = is_scene_visible && camera_choice === 'Main Camera';
       }
     });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [camera_choice, is_scene_visible]);
 
   const handlePlayChange = () => {
@@ -59,16 +58,10 @@ export default function StatusPanel(props: StatusPanelProps) {
       path: 'renderingState/isTraining',
       data: !isTraining,
     });
-    // write to server
-    const cmd = 'write';
-    const path = 'renderingState/isTraining';
-    const data = {
-      type: cmd,
-      path,
-      data: !isTraining,
-    };
-    const message = msgpack.encode(data);
-    websocket.send(message);
+    sendWebsocketMessage(viser_websocket, {
+      type: 'IsTrainingMessage',
+      is_training: !isTraining,
+    });
   };
   const is_training_text = isTraining ? 'Pause Training' : 'Resume Training';
   const training_icon = isTraining ? <PauseIcon /> : <PlayArrowIcon />;
@@ -129,10 +122,10 @@ export default function StatusPanel(props: StatusPanelProps) {
       </Button>
       <div className="StatusPanel-metrics">
         <div>
-          <b>Resolution:</b> {eval_res}
+          <b>Iteration:</b> {step}
         </div>
         <div>
-          <b>Time Allocation:</b> {vis_train_ratio}
+          <b>Resolution:</b> {eval_res}
         </div>
       </div>
       <Button
