@@ -33,13 +33,15 @@ from nerfstudio.pipelines.base_pipeline import Pipeline
 CONSOLE = Console(width=120)
 
 
-def eval_load_checkpoint(config: TrainerConfig, pipeline: Pipeline) -> Path:
+def eval_load_checkpoint(config: TrainerConfig, pipeline: Pipeline) -> Tuple[Path, int]:
     ## TODO: ideally eventually want to get this to be the same as whatever is used to load train checkpoint too
     """Helper function to load checkpointed pipeline
 
     Args:
         config (DictConfig): Configuration of pipeline to load
         pipeline (Pipeline): Pipeline instance of which to load weights
+    Returns:
+        A tuple of the path to the loaded checkpoint and the step at which it was saved.
     """
     assert config.load_dir is not None
     if config.load_step is None:
@@ -61,14 +63,14 @@ def eval_load_checkpoint(config: TrainerConfig, pipeline: Pipeline) -> Path:
     loaded_state = torch.load(load_path, map_location="cpu")
     pipeline.load_pipeline(loaded_state["pipeline"], loaded_state["step"])
     CONSOLE.print(f":white_check_mark: Done loading checkpoint from {load_path}")
-    return load_path
+    return load_path, load_step
 
 
 def eval_setup(
     config_path: Path,
     eval_num_rays_per_chunk: Optional[int] = None,
     test_mode: Literal["test", "val", "inference"] = "test",
-) -> Tuple[TrainerConfig, Pipeline, Path]:
+) -> Tuple[TrainerConfig, Pipeline, Path, int]:
     """Shared setup for loading a saved pipeline for evaluation.
 
     Args:
@@ -81,7 +83,7 @@ def eval_setup(
 
 
     Returns:
-        Loaded config, pipeline module, and corresponding checkpoint.
+        Loaded config, pipeline module, corresponding checkpoint, and step
     """
     # load save config
     config = yaml.load(config_path.read_text(), Loader=yaml.Loader)
@@ -102,6 +104,6 @@ def eval_setup(
     pipeline.eval()
 
     # load checkpointed information
-    checkpoint_path = eval_load_checkpoint(config, pipeline)
+    checkpoint_path, step = eval_load_checkpoint(config, pipeline)
 
-    return config, pipeline, checkpoint_path
+    return config, pipeline, checkpoint_path, step
