@@ -15,12 +15,12 @@
 """Generic utility functions
 """
 
-import sys
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Callable, List, Optional, Tuple
 
 import numpy as np
 import torch
-import zmq
+
+from nerfstudio.viewer.viser.messages import CameraMessage
 
 
 def get_chunks(
@@ -61,7 +61,7 @@ def three_js_perspective_camera_focal_length(fov: float, image_height: int):
 
 
 def get_intrinsics_matrix_and_camera_to_world_h(
-    camera_object: Dict[str, Any], image_height: int, image_width: Optional[int] = None
+    camera_message: CameraMessage, image_height: int, image_width: Optional[int] = None
 ) -> Tuple[torch.Tensor, torch.Tensor]:
     """Returns the camera intrinsics matrix and the camera to world homogeneous matrix.
 
@@ -70,17 +70,17 @@ def get_intrinsics_matrix_and_camera_to_world_h(
         image_size: the size of the image (height, width)
     """
     # intrinsics
-    fov = camera_object["fov"]
-    aspect = camera_object["aspect"]
+    fov = camera_message.fov
+    aspect = camera_message.aspect
     if image_width is None:
         image_width = aspect * image_height
     pp_w = image_width / 2.0
     pp_h = image_height / 2.0
-    if (camera_object["camera_type"] == "perspective") | (camera_object["camera_type"] == "fisheye"):
+    if (camera_message.camera_type == "perspective") | (camera_message.camera_type == "fisheye"):
         focal_length = three_js_perspective_camera_focal_length(fov, image_height)
         intrinsics_matrix = torch.tensor([[focal_length, 0, pp_w], [0, focal_length, pp_h], [0, 0, 1]]).float()
-    elif camera_object["camera_type"] == "equirectangular":
-        render_aspect = camera_object["render_aspect"]
+    elif camera_message.camera_type == "equirectangular":
+        render_aspect = camera_message.render_aspect
         if aspect < render_aspect:
             intrinsics_matrix = torch.tensor(
                 [[pp_w, 0, pp_w], [0, image_width / render_aspect, pp_h], [0, 0, 1]]
@@ -91,7 +91,7 @@ def get_intrinsics_matrix_and_camera_to_world_h(
             ).float()
 
     # extrinsics
-    camera_to_world_h = torch.tensor(get_chunks(camera_object["matrix"], size_of_chunk=4)).T.float()
+    camera_to_world_h = torch.tensor(get_chunks(camera_message.matrix, size_of_chunk=4)).T.float()
     camera_to_world_h = torch.stack(
         [
             camera_to_world_h[0, :],
@@ -117,8 +117,6 @@ def find_available_port(func: Callable, default_port: int, max_attempts: int = 1
         port = default_port + i
         try:
             return func(port, **kwargs), port
-        except (OSError, zmq.error.ZMQError):
-            print(f"Port: {port:d} in use, trying another...", file=sys.stderr)
         except Exception as e:
             print(type(e))
             raise
